Move the swiping tab feature to a common module
