+++ common/common-ui/src/main/java/com/duckduckgo/common/ui/tabs/SwipingTabsFeatureProvider.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2024 DuckDuckGo
+ * Copyright (c) 2025 DuckDuckGo
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.duckduckgo.app.browser
+package com.duckduckgo.common.ui.tabs
 
 import com.duckduckgo.di.scopes.AppScope
 import dagger.SingleInstanceIn
